init.mako.rc: Restart cameraserver when qcamerasvr is starting
As mediaserver has been splitted, there might be a race condition
when starting cameraserver between it and qcamerasvr.

As cameraserver starts sometimes before qcamerasvr, the system
doesn't see a torch available for the Camera and blocks it.

Restart accordingly cameraserver when qcamerasvr is started.

I know that there are better ways of doing this, hoping that
this is a temporary solution.

Thanks to @xyyx for his initial fix:

https://github.com/nitrogen-project/android_device_lge_geehrc/commit/0f728b1f4bd5fe0b2a3e6c8b28fc6baa75714211